backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
sse-starlette>=2.1.0

# LLM & Orchestration - Using compatible versions
langgraph>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.50.0

# RAG & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# File Processing
pypdf>=5.0.0
python-docx>=1.1.0
Pillow>=11.0.0
python-magic>=0.4.27
markdown>=3.7
chardet>=5.2.0
pymupdf>=1.25.0

# Search
ddgs>=1.0.0

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
tenacity>=9.0.0
aiofiles>=24.1.0

# Development
pytest>=8.3.0
httpx>=0.28.0
